refactor(models): fold spell-timing into spells, symmetrize API
spells was missing get/find/list — only findAll/limit for class-scoped
queries. spell-timing was a standalone helper that transformed Spell
arrays but didn't belong outside the model.

- Add spells.get({id}), spells.find({id}), spells.list() for
  per-spell lookup across the full cantrip + level-1 catalog.
- Fold spellCastingTiming → spells.timing(spell) projection.
- Fold groupSpellsByTiming → spells.groupByTiming(spells) projection.
- Delete src/models/spells/spell-timing.ts + its test; port the timing
  cases into spells.test.ts with new coverage for get/find/list.
- Migrate the one consumer (action-bar.tsx).
