+++ src/models/spells/spells.ts
@@ -10,12 +10,13 @@ import warlockSpellsData from "src/data/spells/warlock-spells.json";
 import wizardSpellsData from "src/data/spells/wizard-spells.json";
 
 import type { CharacterClass } from "src/models/class/classes";
+import type { ActionTiming } from "src/models/common/actions";
 import type { DamageType } from "src/models/common/damage";
 
 type SpellId = string;
 
 export type Spell = {
-  id: string;
+  id: SpellId;
   name: string;
   level: number; // 0 = cantrip
   school: string;
@@ -36,6 +37,11 @@ export type Spell = {
 
 const ALL_CANTRIPS = cantripsData as Record<SpellId, Spell>;
 const ALL_SPELLS_LEVEL_1 = spellsLevel1Data as Record<SpellId, Spell>;
+const ALL: Spell[] = [
+  ...Object.values(ALL_CANTRIPS),
+  ...Object.values(ALL_SPELLS_LEVEL_1),
+];
+const BY_ID = new Map(ALL.map((s) => [s.id, s]));
 
 function resolveSpells(data: {
   cantrips: string[];
@@ -66,11 +72,47 @@ const CLASS_DATA: Partial<
   ranger: { spells: resolveSpells(rangerSpellsData), selections: [0, 0] },
 };
 
+function castingTimingOf(spell: Spell): ActionTiming | null {
+  const ct = spell.castingTime.toLowerCase();
+  if (ct === "1 action") return "action";
+  if (ct === "1 bonus action") return "bonus-action";
+  if (ct.startsWith("1 reaction")) return "reaction";
+  return null;
+}
+
 export const spells = {
+  get({ id }: { id: SpellId }): Spell {
+    const found = BY_ID.get(id);
+    if (!found) throw new Error(`Unknown spell: ${id}`);
+    return found;
+  },
+  find({ id }: { id: string }): Spell | undefined {
+    return BY_ID.get(id);
+  },
+  list(): Spell[] {
+    return ALL;
+  },
   findAll({ cls, level }: { cls: CharacterClass; level: 0 | 1 }): Spell[] {
     return CLASS_DATA[cls]?.spells[level] ?? [];
   },
   limit({ cls, level }: { cls: CharacterClass; level: 0 | 1 }): number {
     return CLASS_DATA[cls]?.selections[level] ?? 0;
+  },
+  timing(spell: Spell): ActionTiming | null {
+    return castingTimingOf(spell);
+  },
+  groupByTiming(spellList: Spell[]): Record<ActionTiming, Spell[]> {
+    const groups: Record<ActionTiming, Spell[]> = {
+      action: [],
+      "bonus-action": [],
+      reaction: [],
+    };
+    for (const spell of spellList) {
+      const timing = castingTimingOf(spell);
+      if (timing) {
+        groups[timing].push(spell);
+      }
+    }
+    return groups;
   },
 };
